requirements.txt
yfinance==0.2.55
pandas==2.2.2
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
python-dotenv==1.0.1
apscheduler==3.10.4
requests==2.32.3
